requirements.txt
anndata==0.9.1
byol-pytorch==0.6.0
matplotlib==3.7.1
opencv-python==4.7.0.72
pandas==2.0.1
pot==0.9.0
pyyaml==6.0
rpy2==3.5.11
scanpy==1.9.3
scikit-misc==0.2.0
seaborn==0.12.2
umap-learn==0.5.3
numpy==1.24.3
scikit-learn==1.2.2
scipy==1.10.1
h5py==3.8.0
tqdm==4.65.0

